# src/swagger/swagger.json
{
	"openapi": "3.0.0",
	"components": {
...
		"schemas": {
...
			"ObjectId": {
				"type": "string",
				"description": "A class representation of the BSON ObjectId type."
			},
			"SocialLink": {
				"properties": {
					"platform": {
						"type": "string"
					},
					"url": {
						"type": "string"
					}
				},
				"required": [
					"platform",
					"url"
				],
				"type": "object",
				"additionalProperties": true
			},
			"Pick_User.Exclude_keyofUser.password__": {
				"properties": {
					"name": {
						"type": "string"
					},
					"email": {
						"type": "string"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					},
					"role": {
						"type": "string",
						"enum": [
							"user",
							"admin",
							"superadmin"
						]
					},
					"phone": {
						"type": "string"
					},
					"orders": {
						"items": {
							"$ref": "#/components/schemas/ObjectId"
						},
						"type": "array"
					},
					"addresses": {
						"items": {
							"$ref": "#/components/schemas/ObjectId"
						},
						"type": "array"
					},
					"authorize": {
						"type": "boolean"
					},
					"googleId": {
						"type": "string"
					},
					"facebookId": {
						"type": "string"
					},
					"avatar": {
						"type": "string"
					},
					"profile": {
						"properties": {
							"socialLinks": {
								"items": {
									"$ref": "#/components/schemas/SocialLink"
								},
								"type": "array"
							},
							"description": {
								"type": "string"
							},
							"title": {
								"type": "string"
							},
							"photo": {
								"type": "string"
							}
						},
						"type": "object"
					},
					"_id": {
						"$ref": "#/components/schemas/ObjectId"
					},
					"updatedAt": {
						"type": "string",
						"format": "date-time"
					}
				},
				"required": [
					"name",
					"email",
					"authorize"
				],
				"type": "object",
				"description": "From T, pick a set of properties whose keys are in the union K"
			},
			"Omit_User.password_": {
				"$ref": "#/components/schemas/Pick_User.Exclude_keyofUser.password__",
				"description": "Construct a type with the properties of T except for those in type K."
			},
...
			"ErrorResponse": {
				"properties": {
					"timestamp": {
						"type": "string"
					},
					"error": {
						"properties": {
							"details": {},
							"code": {
								"type": "string"
							},
							"message": {
								"type": "string"
							}
						},
						"required": [
							"message"
						],
						"type": "object"
					},
					"message": {
						"type": "string"
					},
					"success": {
						"type": "boolean",
						"enum": [
							false
						],
						"nullable": false
					}
				},
				"required": [
					"timestamp",
					"error",
					"message",
					"success"
				],
				"type": "object"
			},
...
			"SuccessResponse_Omit_User.password__": {
				"properties": {
					"pagination": {
						"properties": {
							"pages": {
								"type": "number",
								"format": "double"
							},
							"total": {
								"type": "number",
								"format": "double"
							},
							"limit": {
								"type": "number",
								"format": "double"
							},
							"page": {
								"type": "number",
								"format": "double"
							}
						},
						"required": [
							"pages",
							"total",
							"limit",
							"page"
						],
						"type": "object"
					},
					"timestamp": {
						"type": "string"
					},
					"data": {
						"$ref": "#/components/schemas/Omit_User.password_"
					},
					"message": {
						"type": "string"
					},
					"success": {
						"type": "boolean",
						"enum": [
							true
						],
						"nullable": false
					}
				},
				"required": [
					"timestamp",
					"success"
				],
				"type": "object"
			},
			"ApiResponse_Omit_User.password__": {
				"anyOf": [
					{
						"$ref": "#/components/schemas/SuccessResponse_Omit_User.password__"
					},
					{
						"$ref": "#/components/schemas/ErrorResponse"
					}
				]
			},
			"LoginModel": {
				"properties": {
					"userName": {
						"type": "string"
					},
					"userEmail": {
						"type": "string"
					},
					"password": {
						"type": "string"
					}
				},
				"required": [
					"userName",
					"userEmail",
					"password"
				],
				"type": "object",
				"additionalProperties": true
			}
		},
		"securitySchemes": {}
	},
	"info": {
		"title": "app",
		"version": "1.0.0",
		"license": {
			"name": "ISC"
		},
		"contact": {}
	},
	"paths": {
...
		"/api/v1/auth/google/callback": {
			"get": {
				"operationId": "GoogleCallback",
				"responses": {
					"302": {
						"description": "Redirecting"
					}
				},
				"tags": [
					"Auth"
				],
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "code",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
...
		"/api/v1/auth/facebook/callback": {
			"get": {
				"operationId": "FacebookCallback",
				"responses": {
					"302": {
						"description": "Redirecting"
					}
				},
				"tags": [
					"Auth"
				],
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "code",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
...
		"/api/v1/auth/login": {
			"post": {
				"operationId": "LoginUser",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_Omit_User.password__"
								}
							}
						}
					}
				},
				"tags": [
					"Auth"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/LoginModel"
							}
						}
					}
				}
			}
		},
...
		"/api/v1/auth/forgot-password": {
			"post": {
				"operationId": "ResetPassword",
				"responses": {
					"200": {
						"description": "Reset Link Sent",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"error": {}
									},
									"type": "object"
								}
							}
						}
					}
				},
				"tags": [
					"Auth"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"email": {
										"type": "string"
									}
								},
								"required": [
									"email"
								],
								"type": "object"
							}
						}
					}
				}
			}
		},
		"/api/v1/auth/verify-reset-token": {
			"get": {
				"operationId": "VerifyResetToken",
				"responses": {
					"200": {
						"description": "Token Valid",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"tags": [
					"Auth"
				],
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "token",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/api/v1/auth/reset-password": {
			"post": {
				"operationId": "SetNewPassword",
				"responses": {
					"200": {
						"description": "Password Reset Successfully",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"tags": [
					"Auth"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"newPassword": {
										"type": "string"
									},
									"token": {
										"type": "string"
									}
								},
								"required": [
									"newPassword",
									"token"
								],
								"type": "object"
							}
						}
					}
				}
			}
		}
	},
	"servers": [
		{
			"url": "/"
		}
	]
}